--- task_checker_back/taskchecker-teamdevelopm-6580e-firebase-adminsdk-fbsvc-aad758dfec.jsonZone.Identifier
@@ -1,3 +1,0 @@
-[ZoneTransfer]
-ZoneId=3
-HostUrl=about:internet
